Use the same set of AMIs for uploading, regardless of base system.
We just need an EC2 instance that we can SSH into and scp the image
over. It also needs to have dd_rescue installed in order for us to
efficiently dump the image to the EBS volume. The SLES11 SP2 public AMIs
satisfy all these requirements, so let's just use that everywhere.